Move TcgaCnvParser to its own file under input_data_adapters.  Adjust main() and tests accordingly.
Additionally, some extra dataframe copies were added to TcgaCnvParser.get_top_genes(), which continues to have maddening flaky segmentation faults.  It was hoped that more explicit semantics for dataframe types would mitigate this, but alas any benefit seems to be a minimal one.
